+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - EXT-X-SESSION-KEY support
 - EXT-X-CONTENT-STEERING support
+- EXT-X-START to master playlist
+
+### Fixed
+
+- EXT-X-START was not written for negative values
 
 ## [v0.3.0] 2025-01-14